test: deduplicate barrier helper
